[fix] : #26 CSV관련 위치 및 rename
