+++ src/main/java/ussum/homepage/infra/csvBatch/csv/CSVReader.java
@@ -8,24 +8,23 @@ import org.springframework.batch.item.file.transform.DelimitedLineTokenizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.io.ClassPathResource;
-import ussum.homepage.infra.csvBatch.csv.entity.StudentCsvDto;
+import ussum.homepage.domain.csv_user.StudentCsv;
 
 @Configuration
 @RequiredArgsConstructor
 public class CSVReader {
     @Bean
-    public FlatFileItemReader<StudentCsvDto> csvFileItemReader() { // 엔티티를 반환하는게 맞나?
+    public FlatFileItemReader<StudentCsv> csvFileItemReader() { // 엔티티를 반환하는게 맞나?
         /* file read */
-        FlatFileItemReader<StudentCsvDto> flatFileItemReader = new FlatFileItemReader<>();
+        FlatFileItemReader<StudentCsv> flatFileItemReader = new FlatFileItemReader<>();
         flatFileItemReader.setResource(new ClassPathResource("csv/학생목록.csv"));
         flatFileItemReader.setLinesToSkip(1); // header line skip
         flatFileItemReader.setEncoding("UTF-8"); // encoding
 //        flatFileItemReader.setStrict(false);
         // flatFileItemReader.setRecordSeparatorPolicy(new DefaultRecordSeparatorPolicy());
-        // 개행이 있을 시 사용
 
         /* read하는 데이터를 내부적으로 LineMapper을 통해 Mapping */
-        DefaultLineMapper<StudentCsvDto> defaultLineMapper = new DefaultLineMapper<>();
+        DefaultLineMapper<StudentCsv> defaultLineMapper = new DefaultLineMapper<>();
 
         /* delimitedLineTokenizer : setNames를 통해 각각의 데이터의 이름 설정 */
         DelimitedLineTokenizer delimitedLineTokenizer = new DelimitedLineTokenizer(",");
@@ -42,8 +41,8 @@ public class CSVReader {
         defaultLineMapper.setLineTokenizer(delimitedLineTokenizer);
 
         /* beanWrapperFieldSetMapper : Tokenizer에서 가지고온 데이터들을 VO로 바인드하는 역할 */
-        BeanWrapperFieldSetMapper<StudentCsvDto> beanWrapperFieldSetMapper = new BeanWrapperFieldSetMapper<>();
-        beanWrapperFieldSetMapper.setTargetType(StudentCsvDto.class);
+        BeanWrapperFieldSetMapper<StudentCsv> beanWrapperFieldSetMapper = new BeanWrapperFieldSetMapper<>();
+        beanWrapperFieldSetMapper.setTargetType(StudentCsv.class);
 
         defaultLineMapper.setFieldSetMapper(beanWrapperFieldSetMapper);
