+++ src/main/java/ussum/homepage/infra/csvBatch/csv/CSVWriter.java
@@ -5,24 +5,24 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.batch.item.Chunk;
 import org.springframework.batch.item.ItemWriter;
 import org.springframework.context.annotation.Configuration;
-import ussum.homepage.infra.csvBatch.csv.entity.StudentCsvData;
-import ussum.homepage.infra.csvBatch.csv.entity.StudentCsvDto;
-import ussum.homepage.infra.csvBatch.csv.repository.StudentCsvRepository;
+import ussum.homepage.domain.csv_user.StudentCsv;
+import ussum.homepage.domain.csv_user.StudentCsvRepository;
+import ussum.homepage.infra.jpa.csv_user.entity.StudentCsvEntity;
 
 @Configuration
 @RequiredArgsConstructor
 @Transactional
-public class CSVWriter implements ItemWriter<StudentCsvDto> {
+public class CSVWriter implements ItemWriter<StudentCsv> {
 
     private final StudentCsvRepository csvRepository;
 
     @Override
     @Transactional
-    public void write(Chunk<? extends StudentCsvDto> chunk) throws Exception {
-        Chunk<StudentCsvData> students = new Chunk<>();
+    public void write(Chunk<? extends StudentCsv> chunk) throws Exception {
+        Chunk<StudentCsvEntity> students = new Chunk<>();
 
         chunk.forEach(studentCsvDto -> {
-            StudentCsvData data = StudentCsvData.of(studentCsvDto.getSTID(), studentCsvDto.getStudentId(), studentCsvDto.getStudentName(), studentCsvDto.getGroupName(),
+            StudentCsvEntity data = ussum.homepage.infra.jpa.csv_user.entity.StudentCsvEntity.of(studentCsvDto.getSTID(), studentCsvDto.getStudentId(), studentCsvDto.getStudentName(), studentCsvDto.getGroupName(),
                     studentCsvDto.getMajor(), studentCsvDto.getStudentStatus(), studentCsvDto.getStudentGroup(), studentCsvDto.getStudentEmail());
             students.add(data);
         });
